DateFormatter movido a la vista, determina si es fecha invalida
